requirements.txt
Flask
python-dotenv
psycopg2-binary
requests
gunicorn
bcrypt
